+++ src/edt/edt/edtPropertiesPages.cc
@@ -73,7 +73,7 @@ ShapePropertiesPage::setup ()
 void 
 ShapePropertiesPage::back ()
 {
-  m_index = m_selection_ptrs.size ();
+  m_index = (unsigned int) m_selection_ptrs.size ();
 }
 
 void
